Задание 7, внесла изменения
